Move mapview layer stuff out of tilespec.h
The map drawing code will be modularized and layers can ~easily be abstracted
away.

See #430.
